src/scripts/video/requirements.txt
PyVimeo>=1.1.2 